Improve perf by favoring selection of division games
This assumes the requirements for division games are harder to meet
than for non-divisional games, so it favors resolving those matchups
first. The side effect of this combined with always searching down the
path of least options is that all division games (across all divisions)
will always occur on the same weeks. This isn't ideal, and hopefully
won't be the final behavior.

This commit also removes the fixed random seed that was used for
testing.